Fixed the variable binding name test.
It failed because showing the variable names of the posting knowledge
base within the exchange infos is no longer possible when we use the
reasoner because there might be rules between the poster and some
reacting knowledge base. In that case using the variable names from the
poster does not make any sense.