Fixed the errors in the test.
The TripleVarBindingSet#translate() was incorrect, if one match was
skipped, all of them were.

Also added some convenient methods for debugging in the bindingset
classes.

The TestAskAnswerKBMulitpleMatchingKIs test failed once, but then it
succeeded again. This was caused by there only being 2 KBs involved in
its ask instead of the expected 3. When tested again, it succeeded and
it kept succeeding after that. So, there is still something wrong with
this code, but it only happens sometimes. Probably something to do with
the meta messages and they not all being processed correctly.