Moved Knowledge Gap logic to smart-connector module.
Tested two types of knowledge gaps: the first is a knowledge gap in
which two triple patterns need to become available for the gaps to be
filled, while the second is a knowledge gap in which either one or
another triple needs to be provided to fill the gap. So, AND versus OR.
You can choose between them by commenting out the ruleSet#add(...)
calls.


Here we can ignore meta knowledge interactions and prevent them from
removing many gaps. Within the reasoner module we could not do this,
because we did not have access to the smart-connector module and we want
to keep it that way!