Make sure we reuse the KeRuntime threadpool.
Both the http client and server (that are being used when a Knowledge
Directory is available) initialize their own threadpool that prevent
Java applications from shutting down (because we actually should
shutdown the Jetty server as well). This commit changes that so that
those client and server use the KeRuntime threadpool.