Implement a timeout.
We need a timeout after all. Does this interfere with our strive to
support human-in-the-loop and some KBs might take days to answer? I do
not think this interferes, because the timeout is on sending the REST
message and receiving a 200. The timeout for human-in-the-loop is on how
long the SerialMatchingProcessor will wait until it continues with
processing the next matching KI. Here we want to be able to wait
indefinitely.