Add max capacity on maps to prevent memory leaks.
Done:
- Two maps in MessageRouterImpl now remove (and cancel) oldest entry
when max capacity has been reached.
- beingProcessed in RestKnowledgeBase now also removes) and cancels)
oldest eentry when max capacity has been reached.
- Improve error handling of SerialMatchingProcessor when max capacity
has been reached.

The RestKnowledgeBase max capacity is much lower than that of the
MessageRouterImpl, because we expect the RestKnowledgeBase to correctly
deal with too many handle request (and this prevents the
messageRouterImpl maps to reach max capacity).