Fix ordering issue when starting two runtimes.
Issue was:
- when we first start our communication, other KERs do not know our KBs
yet when we send them a message to announce ourselves.
- when we first let other KERs know, our communication is not yet
started when we receive (and want to reply) to messages from other KERs.

The current version is that we do not immediately send a message to
another KER when it contacts us, but we wait for it to announce itself.
However, not sure whether this actually works in all scenario's, so we
need to test this further or think about it further.