First version of new algorithm.
Had to introduce new strategy that limits the full matching to a single
rule. This was neccesary to make meta-knowledge-interactions work
faster.

Also had to remove some invert match code, which I found strange, but
had to do to make it work.

There are still quite some tests the fail, so those need to be fixed
first. Starting with the looping test.