All functionality of the matcher is there.
However, it does not fully work yet. Some tests are still failing and we
need to compare whether the results are comparable with the previous
graph pattern matcher. Also, maybe discuss the different levels of
interoperability with others and make sure the TKE fits into that.

Extra configuration options to add to TKE (as opposed to reasoner):
- do not include meta graph patterns (ReasonerProcessor)
- only go one level deep with reasoning
(ReasoningPlan#createOrGetReasoningNode)