Slightly adapt unit test to test additional scenario.
This new unit test would have failed in version 1.2.4, but is fixed in
version 1.3.0. The problem is an ASK graph pattern with two variables,
while the ANSWER graph pattern has one of these variables filled in with
a concrete value. This concrete value would not end up in the result,
which is wrong.